Former-commit-id: a486dd4451c954b3113bb54491bd0da2493bb3b6

+++ arch/x86/kernel/head64.c
@@ -96,7 +96,7 @@ void __init x86_64_start_kernel(char * real_mode_data)
 	if (console_loglevel == 10)
 		early_printk("Kernel alive\n");
 
-	orig_boot_params = real_mode_data;
+	orig_boot_params = (unsigned long) real_mode_data;
 
 	x86_64_start_reservations(real_mode_data);
 }
